refactor code using modules
