+++ js/game/reloadBtn.js
@@ -1,0 +1,17 @@
+import { createElement } from "../common/common.js";
+import { rootEl } from "./player.js";
+
+const createReloadButton = () => {
+  const reloadBtnWrapEl = createElement('div', 'reloadWrap');
+  const reloadBtnEl = createElement('button', 'button');
+
+  reloadBtnEl.innerText = 'Restart';
+  reloadBtnWrapEl.appendChild(reloadBtnEl);
+  rootEl.appendChild(reloadBtnWrapEl);
+
+  reloadBtnEl.addEventListener('click', function () {
+    window.location.reload();
+  });
+}
+
+export { createReloadButton }
